salesproject/requirements.txt
streamlit as st
plotly.express as px
pandas as pd
matplotlib.pyplot as plt
os
seaborn as sns
warnings
warnings.filterwarnings('ignore')
statsmodels.api as sm